Update lyrics.R to handle instrumental songs
Previously, scrape_lyrics_url() would throw an error when it encountered an instrumental song. This checks to see if there are no lyrics left after removing items in square brackets, and if so, creates a single item with an NA value so that it can be converted to a tibble. This change resolved the error I was having with instrumental songs.